backend/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
asyncpg==0.30.0
cbor2==5.6.5
cffi==1.17.1
click==8.2.1
cryptography==45.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
webauthn==2.5.2
websockets==15.0.1
